refactor(cli): extract parseCliArgs into its own module
Move parseCliArgs and the local parsePortValue helper out of
src/cli/index.ts into a dedicated src/cli/parse-args.ts so the
bootstrap test can import them without dragging in the whole CLI
graph — HttpServer, SecretsStore, initPersistenceRuntime, the
services wiring and their transitives.

After ~3700 unit tests have warmed Vitest's module cache, loading
that graph on the test's dynamic import had started to exceed the
default 5s per-test timeout, flaking the pre-push gate. The prior
"stabilize" commit switched from main() to parseCliArgs() but left
the import path wide; this narrows it to just node:util + the
logger init, which the test actually needs.

src/cli/index.ts re-exports parseCliArgs so the existing integration
test and any other consumers keep working without changes.
